delete saptie library and recreate user management CRUD

+++ app/Http/Controllers/UserController.php
@@ -13,11 +13,10 @@ class UserController extends Controller
 
     public function index(Request $request) {
         $users = User::withTrashed()
-            ->with('roles')
-            ->select(['id', 'name', 'email', 'deleted_at'])
+            ->with('role')
+            ->select(['id', 'name', 'email', 'deleted_at', 'role_id'])
             ->paginate(25);
-        
-        $roles = Role::all();
+            $roles = Role::all();
         return view('pages.dashboard.user.index', compact('roles', 'users'));
     }
 
@@ -31,34 +30,22 @@ class UserController extends Controller
             'name' => 'required|string|max:255',
             'email' => 'required|string|email|max:255|unique:users,email',
             'password' => 'required|min:8',
-            'role' => 'required|integer|in:1,2,3',
+            'role' => 'required|exists:roles,id',
         ]);
-
+        
         $user = User::create([
             'name' => $validatedData['name'],
             'email' => $validatedData['email'],
             'password' => Hash::make($validatedData['password']),
+            'role_id' => (int) $validatedData['role'], // Cast 'role' to integer
         ]);
-
-        // Assign role based on the number
-        switch ($validatedData['role']) {
-            case 1:
-                $user->assignRole('manager');
-                break;
-            case 2:
-                $user->assignRole('admin');
-                break;
-            case 3:
-                $user->assignRole('employee');
-                break;
-        }
 
         return redirect()->route('users.index')->with('success', 'User created successfully');
     }
 
     public function edit($id) {
         $roles = Role::all();
-        $user = User::withTrashed()->with('roles')->find($id);
+        $user = User::withTrashed()->with('role')->find($id);
 
         if (!$user) {
             return redirect()->route('users.index')->with('error', 'User not found.');
@@ -82,50 +69,34 @@ class UserController extends Controller
         // Update user details
         $user->name = $validatedData['name'];
         $user->email = $validatedData['email'];
+        $user->role_id = $validatedData['role'];
         if (!empty($validatedData['password'])) {
             $user->password = bcrypt($validatedData['password']);
         }
         $user->save();
 
-        // Remove all roles before assigning the new one
-        $user->roles()->detach();
-
-        // Assign role based on the number
-        switch ($validatedData['role']) {
-            case 1:
-                $user->assignRole('manager');
-                break;
-            case 2:
-                $user->assignRole('admin');
-                break;
-            case 3:
-                $user->assignRole('employee');
-                break;
-        }
 
         return redirect()->route('users.index')->with('success', 'User updated successfully');
     }
 
     public function suspend(User $user) {
+        if($user->role_id === 1) {
+            return redirect()->route('users.index')->with('error', 'Can\'t suspend Manager Role!');
+        }
         $user->delete();
         return redirect()->route('users.index')->with('success', 'User suspended successfully');
     }
 
     public function restore($id) {
         $user = User::withTrashed()->find($id);
-
+        if($user->role_id === 1) {
+            return redirect()->route('users.index')->with('error', 'Can\'t suspend Manager Role!');
+        }
         if ($user && $user->trashed()) {
             $user->restore();
             return redirect()->route('users.index')->with('success', 'User restored successfully.');
         }
 
         return redirect()->route('users.index')->with('error', 'User not found or not deleted.');
-    }
-
-
-
-    public function assignRole(User $user, Request $request){
-        $user->assignRole($request->role);
-        return redirect()->route('users.index')->with('success', 'Role assigned successfully');
     }
 }
